+++ app/controllers/report.js
@@ -61,28 +61,6 @@ export default Ember.Controller.extend({
     handleResize(e) {
       this.fitBounds(e.target);
     },
-
-    handleDownload(format = 'csv') {
-      const report = this.get('model').sortBy('year', 'profile', 'category').reverse();
-      const columnNames = [Object.keys(report.get('firstObject'))];
-      const matrixValues = report.map(
-        row => Object.values(row),
-      );
-
-      const data = []
-        .concat(
-          columnNames,
-          matrixValues,
-        );
-
-      if (format === 'csv') {
-        this.get('csv').export(data, { fileName: 'download.csv', withSeparator: false });
-      }
-
-      if (format === 'excel') {
-        this.get('excel').export(data, { sheetName: 'sheet1', fileName: 'download.xlsx' });
-      }
-    },
   },
 
 });
