+++ app/controllers/report/demographic.js
@@ -63,31 +63,4 @@ export default Ember.Controller.extend({
       pyramidData,
     };
   },
-
-  actions: {
-    handleCopy() {
-      const el = document.getElementsByTagName('table')[0];
-      const body = document.body;
-      let range;
-      let sel;
-      if (document.createRange && window.getSelection) {
-        range = document.createRange();
-        sel = window.getSelection();
-        sel.removeAllRanges();
-        try {
-          range.selectNodeContents(el);
-          sel.addRange(range);
-        } catch (e) {
-          range.selectNode(el);
-          sel.addRange(range);
-        }
-      } else if (body.createTextRange) {
-        range = body.createTextRange();
-        range.moveToElementText(el);
-        range.select();
-      }
-
-      document.execCommand('copy');
-    },
-  },
 });
