+++ app/routes/report/demographic.js
@@ -1,7 +1,8 @@
 import Ember from 'ember';
 import carto from 'ember-jane-maps/utils/carto';
-import nestReport from '../../utils/nest-report';
 import generateReportSQL from '../../queries/report';
+
+import nestReport from '../../utils/nest-report';
 
 const { service } = Ember.inject;
 
@@ -11,7 +12,14 @@ export default Ember.Route.extend({
     const geoids = this.get('selection.current.features').mapBy('properties.geoid');
     const selectionSQL = generateReportSQL(geoids, comparator, 'demographic');
 
-    return carto.SQL(selectionSQL, 'json', 'post')
-      .then(data => nestReport(data));
+    return carto.SQL(selectionSQL, 'json', 'post');
+  },
+
+  setupController(controller, model) {
+    this._super(controller, model);
+    controller.setProperties({
+      model: nestReport(model),
+      rawData: model,
+    });
   },
 });
