+++ app/templates/components/acs-table.hbs
@@ -1,4 +1,4 @@
-
+<a class="button small expanded hollow" {{action 'handleCopy'}}>Copy to Clipboard</a>
 <div class="table-scroll">
   <table class="text-small {{if options.scrollTable 'table-scroll-vert'}}">
     <thead>
