+++ app/templates/components/download-spreadsheet.hbs
@@ -1,0 +1,3 @@
+<a class='button small expanded hollow' {{action 'handleDownload'}}>
+  Download this Profile (csv)
+</a>
