+++ app/templates/report.hbs
@@ -23,14 +23,8 @@
       </ul>
     </div>
 
-    <a class='button small expanded hollow' {{action 'handleDownload'}}>
-      Download this Profile (csv)
-    </a>
-
   </div>
   <div class="cell auto">
-
     {{outlet}}
-
   </div>
 </div>
