+++ app/templates/report/demographic.hbs
@@ -1,5 +1,4 @@
-
-
+{{download-spreadsheet data=rawData}}
 {{#report-header
   title=(if (eq report.mode 'current') "Demographic Profile" "Demographic Profile of Change Over Time")
   subtitle=(if (eq report.mode 'current') "American Community Survey (ACS), 2011-2015" "American Community Survey (ACS), 2006-2010 to 2011-2015")
@@ -33,7 +32,6 @@
 <h3>Age and Sex</h3>
 <div class="grid-x grid-margin-x">
   <div class="cell large-8">
-    <a class="button small expanded hollow" {{action 'handleCopy'}}>Copy to Clipboard</a>
     {{#acs-table options=report}}
       {{#each sexAge as |row|}}
         {{acs-table-row
